feat: add canonical mine cli and indexer chunker
